Authorization done

+++ app/Http/Controllers/Admin/UserController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers\Admin;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
+
 use App\User;
 
 class UserController extends Controller
@@ -26,7 +28,11 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view('admin.user.create');
+        if (Gate::allows('admin',auth()->user())) {
+            return view('admin.user.create');
+        }
+        session()->flash('message','You are not authorize this operation!!');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -37,25 +43,30 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required',
-            'email' => 'required|email|unique:users',
-            'password' => 'required|confirmed',
-            'image' => 'mimes:jpeg,png,webp'
-        ]);
-        $data = $request->except(['_token','password','image']);
-        $data['password'] = bcrypt($request->password);
-        if ($request->hasfile('image')) {
-            $file = $request->file('image');
-            $path = 'images/user';
-            $file_name = time().rand(0000,9999).'.'.$file->getClientOriginalExtension();
-            $file->move($path,$file_name);
-            $data['image'] = $path.'/'.$file_name;
+
+        if (Gate::allows('admin',auth()->user())) {
+            $request->validate([
+                'name' => 'required',
+                'email' => 'required|email|unique:users',
+                'password' => 'required|confirmed',
+                'image' => 'mimes:jpeg,png,webp'
+            ]);
+            $data = $request->except(['_token','password','image']);
+            $data['password'] = bcrypt($request->password);
+            if ($request->hasfile('image')) {
+                $file = $request->file('image');
+                $path = 'images/user';
+                $file_name = time().rand(0000,9999).'.'.$file->getClientOriginalExtension();
+                $file->move($path,$file_name);
+                $data['image'] = $path.'/'.$file_name;
+            }
+            // dd($request->all());
+            User::create($data);
+            session()->flash('message','Admin create successfully');
+            return redirect()->route('user.index');
         }
-        // dd($request->all());
-        User::create($data);
-        session()->flash('message','Admin create successfully');
-        return redirect()->route('user.index');
+        session()->flash('message','You are not authorize this operation!!');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -77,8 +88,13 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $data['user'] = User::findOrFail($id);
-        return view('admin.user.edit',$data);
+
+        if (Gate::allows('admin',auth()->user())) {
+            $data['user'] = User::findOrFail($id);
+            return view('admin.user.edit',$data);
+        }
+        session()->flash('message','You are not authorize this operation!!');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -90,26 +106,30 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'name' => 'required',
-            'image' => 'mimes:jpeg,png,webp'
-        ]);
-        $data['name'] = $request->name;
-        $user = User::findOrFail($id);
-        if ($request->hasfile('image')) {
-            $file = $request->file('image');
-            $path = 'images/user';
-            $file_name = time().rand(0000,9999).'.'.$file->getClientOriginalExtension();
-            $file->move($path,$file_name);
-            $data['image'] = $path.'/'.$file_name;
+        if (Gate::allows('admin',auth()->user())) {
+            $request->validate([
+                'name' => 'required',
+                'image' => 'mimes:jpeg,png,webp'
+            ]);
+            $data['name'] = $request->name;
+            $user = User::findOrFail($id);
+            if ($request->hasfile('image')) {
+                $file = $request->file('image');
+                $path = 'images/user';
+                $file_name = time().rand(0000,9999).'.'.$file->getClientOriginalExtension();
+                $file->move($path,$file_name);
+                $data['image'] = $path.'/'.$file_name;
+            }
+            if ($user->image != null && file_exists($user->image)) {
+                unlink($user->image);
+            }
+            // dd($request->all());
+            $user->update($data);
+            session()->flash('message','Admin Update successfully');
+            return redirect()->route('user.index');
         }
-        if ($user->image != null && file_exists($user->image)) {
-            unlink($user->image);
-        }
-        // dd($request->all());
-        $user->update($data);
-        session()->flash('message','Admin Update successfully');
-        return redirect()->route('user.index');
+        session()->flash('message','You are not authorize this operation!!');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -120,9 +140,13 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        if (Gate::allows('admin',auth()->user())) {
         //
-        User::destroy($id);
-        session()->flash('message','Delete successfully');
-        return redirect()->back();
+            User::destroy($id);
+            session()->flash('message','Delete successfully');
+            return redirect()->back();
+        }
+        session()->flash('message','You are not authorize this operation!!');
+        return redirect()->route('admin.dashboard');
     }
 }
